refactor: move the main activity screen to new file
Split the MainActivity code up into a few different files to keep it tidy.
